requirements.txt
amqp==1.4.6
anyjson==0.3.3
app-version==0.2.1
billiard==3.3.0.20
boto3==1.1.4
botocore==1.2.7
celery==3.1.18
Django==1.8.4
django-appconf==1.0.1
django-celery==3.1.16
django-debug-toolbar==1.3.2
django-guardian==1.3
django-loginas==0.1.5
django-modeltranslation==0.10.1
django-mptt==0.7.4
django-polymorphic==0.7.2
django-polymorphic-tree==1.1.1
django-recaptcha==1.0.4
django-redis==4.2.0
django-registration-redux==1.2
django-tag-parser==2.1
django-tools==0.29.5
django-uuslug==1.1.6
docutils==0.12
elasticsearch==1.7.0
elasticsearch-dsl==0.0.8
future==0.15.0
futures==2.2.0
google-api-python-client==1.4.2
httplib2==0.9.1
jmespath==0.9.0
kombu==3.0.26
lxml==3.4.4
msgpack-python==0.4.6
oauth2client==1.5.1
Pillow==3.0.0
pip-autoremove==0.9.0
psycopg2==2.6.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pygeoip==0.3.2
python-dateutil==2.4.2
python-slugify==1.1.4
pytz==2015.6
redis==2.10.3
requests==2.7.0
rsa==3.2
simplejson==3.8.0
six==1.9.0
sockjs-tornado==1.0.2
sqlparse==0.1.16
tinys3==0.1.11
tolerance==0.1.2
tornado==4.2.1
tornado-redis==2.4.18
Unidecode==0.4.18
uritemplate==0.6
urllib3==1.12
